refactor: SGDP optimizer
